refactor: Consolidate error handling into single file (TODO_165)
Consolidated fragmented error handling from 3 files into 1 streamlined
implementation, reducing complexity and improving maintainability.

Changes:
- Simplified errors.ts from 126 LOC to 114 LOC
- Deleted error-sanitizer.ts (87 LOC)
- Deleted error-helpers.ts (29 LOC)
- Total reduction: 128 LOC (53%)

Implementation:
- Moved getErrorMessage() from error-helpers.ts to errors.ts
- Moved getErrorStatus() from error-sanitizer.ts to errors.ts
- Updated api-response.ts to use consolidated utilities
- Preserved ValidationError and AppError (actually used in 5 places)
- Removed 7 unused custom error classes (AuthenticationError, etc.)
- Updated popularity-tracker.ts import path
- Removed obsolete tests for deleted sanitization functions
- Archived deleted files to todos/archive/code-removed/

Benefits:
- Single source of truth for error handling
- Reduced cognitive load (1 file vs 3)
- Easier maintenance and debugging
- No functionality loss
- All tests passing (20/20)

See: todos/TODO_165_ERROR_HANDLING_CONSOLIDATION.md
Source: GitHub Issue #165
