+++ .eslintignore
@@ -1,3 +1,4 @@
+
 # Dependencies
 node_modules/
 .pnpm-store/
