+++ todos/archive/code-removed/2025-12-05-error-helpers.ts
